+++ src/Commands/LaravelStrapiCommand.php
@@ -6,12 +6,12 @@ use Illuminate\Console\Command;
 
 class LaravelStrapiCommand extends Command
 {
-    public $signature = 'laravel-strapi';
+    public $signature = 'strapi {task}';
 
     public $description = 'Laravel Strapi Helper';
 
     public function handle()
     {
-        $this->comment('All done');
+        //
     }
 }