+++ src/Exceptions/UnknownError.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Dbfx\LaravelStrapi\Exceptions;
+
+use RuntimeException;
+
+class UnknownError extends RuntimeException
+{
+}